Subdividable objects.  Subdivision objects.
Subdivision objects reduce themselves, e.g. a tetragonal (3D)
subdivision will reduce to a trigonal (2D) subdivision if the
subdivision point lies on a face of the tetrahedron.  Only the
ineviatable subdivisions are made.

The objects (Point, Line, Face, Tetrahedron) support subdivision
by creation of new objects and replacement of existing ones with
a list of new, similar objects.

The inheritance mechanism to define Renderers for all objects is
implemented.  Renderers are given only to Points at initialisation
time or later, all other objects inherit the renderers from the
points they are attached to, if the resp. renderer is attached to
all of the constituting points.

Some other, minor changes.