Adding and integrating the World class.
The World class can initialise from a cuboid.

It has been avoided to store the world in all objects, first because
this is a reference loop, and second to conserve memory.  Instead
the world object is passed in or is available as attribute of arguments
passed in when the world may change.