First version of the automatic subdivision implemented.
The algorithm is untested.  There may be numerous typos and mistakes.

One creates a World instance with finite extent, and creates Points
by a method of this World instance.  Then Lines, Faces etc. can be
created by the normal API, and will in their __init__ subdivide
themselves according to their neighbourhood.

Points outside the world are created "invisible", i.e., all objects
sharing them will be invisible too.  By this method, one can clip
faces etc. at the boundary of the world.  Only the inner, subdivided
parts of the face or similar will appear rendered.  As Renderers are
applies as usual to the invisible objects, they are passed on to the
inner, visible parts as usual.